Fix addUserToClickTime
